move test data to sources folders
