+++ tests/unstructured_data/test_data/invoice_2.txt
@@ -1,0 +1,31 @@
+Invoice Description: Professional Photography and Post-Production Services
+
+Invoice Number: INV-231789
+
+Invoice Date: June 30, 2023
+
+Billed To:
+Customer Name: 123 Company
+Address: 789 Elm Street, Cityville, State
+
+Item Description:
+
+Pre-Production Planning (2 hours): Our team collaborated with the client to develop a comprehensive plan for the photography session. We discussed their vision, desired outcomes, and specific requirements to ensure a successful shoot.
+
+On-Location Photography (4 hours): Our skilled photographers conducted an on-location photo shoot, capturing high-quality images that showcase the client's products, services, or desired subjects. We utilized professional-grade equipment and lighting techniques to achieve stunning visuals.
+
+Post-Production Editing: Our expert editors meticulously processed and retouched the captured photographs to enhance their visual appeal. We employed advanced editing techniques to adjust lighting, colors, and composition, resulting in polished and professional images.
+
+Image Selection and Delivery: We carefully reviewed and selected the best photographs from the session, considering the client's preferences and project objectives. The finalized images were then delivered in digital format, ensuring convenient access and usability.
+
+Subtotal: $1,800.00
+Tax (3%): $54.00
+Total Amount: $1,854.00
+
+Payment Terms: Payment is due within 14 days of the invoice date.
+
+We appreciate the opportunity to provide our professional photography and post-production services. If you have any questions or require further assistance, please feel free to reach out. We value your business and strive to exceed your expectations with our exceptional imagery.
+
+
+
+
