+++ tests/unstructured_data/test_data/invoice_3.jpg
@@ -1,0 +1,27 @@
+Invoice Description: Premium Widget Delivery and Installation Services
+
+Invoice Number: INV-549283
+
+Invoice Date: June 30, 2023
+
+Billed To:
+Customer Name: XYZ Corporation
+Address: 123 Main Street, Cityville, State
+
+Item Description:
+
+Premium Widgets (Qty: 50): These high-quality widgets are manufactured using state-of-the-art technology, ensuring durability and superior performance. Each widget comes with advanced features and a sleek design.
+
+Delivery Service: Our team provided efficient and secure delivery of the premium widgets to the customer's specified location. We ensured proper handling and packaging to prevent any damage during transportation.
+
+Installation Service: Our skilled technicians conducted professional installation of the widgets, ensuring they were seamlessly integrated into the customer's existing system. They followed industry best practices and safety guidelines to guarantee optimal functionality.
+
+Technical Support (3 hours): We offered comprehensive technical support to the customer, assisting with any queries or issues related to the widgets or their installation. Our knowledgeable support team provided prompt and reliable assistance, ensuring smooth operations.
+
+Subtotal: $10,500.00
+Tax (7%): $735.00
+Total Amount: $11,235.00
+
+Payment Terms: Payment is due within 30 days of the invoice date.
+
+Thank you for choosing our premium widget delivery and installation services. We appreciate your business and look forward to serving you in the future.
